pyproject.toml
[tool.poetry]
name = "agentic-ai"
version = "1.0.0"
description = "Production-grade Agentic AI platform with RAG, multi-tenant SaaS, and financial intelligence"
authors = ["vihaan <dev@local>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"
python-dotenv = "^1.0.1"
structlog = "^24.4.0"
httpx = "^0.27.0"
langgraph = "^0.2.0"
langchain-core = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-text-splitters = "^0.3.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.30.0"
psycopg = {version = "^3.2.0", extras = ["binary"]}
alembic = "^1.13.0"
pgvector = "^0.3.0"
python-multipart = "^0.0.9"
pypdf = "^4.3.0"
python-docx = "^1.1.0"
redis = { extras = ["hiredis"], version = "^5.0.0" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
bcrypt = "4.1.3"
authlib = "^1.3.0"
email-validator = "^2.2.0"
celery = { version = "^5.4.0", extras = ["redis"] }
stripe = "^10.0.0"
opentelemetry-api = "^1.27.0"
opentelemetry-sdk = "^1.27.0"
opentelemetry-instrumentation-fastapi = "^0.48b0"
opentelemetry-exporter-otlp = "^1.27.0"
prometheus-client = "^0.21.0"
cryptography = "^43.0.0"
sse-starlette = "^2.1.0"
pyjwt = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
